==> 1-downloader.ipynb <==
import yfinance as yf
import pandas as pd


# Download EUR/USD data
data = yf.download('USDINR=X', 
                     start='2025-01-01', 
                     end='2025-09-01',
                     interval='1d')
data.to_csv('Data/USDINR_day_2025-01-2025-09-Infer.csv')

# Available intervals: 1m, 2m, 5m, 15m, 30m, 60m, 90m, 1h, 1d, 5d, 1wk, 1mo, 3mo


data.head()
